Resolve branch conflicts: consolidate Dispatcher implementations and remove duplicate file
